+++ src/lib/transcriber/TranscriberResult.dart
