+++ src/lib/transcriber/TranscriberSpeechToText.dart
@@ -1,4 +1,4 @@
-import 'package:com_4_all/TranscriberResult.dart';
+import 'package:com_4_all/transcriber/TranscriberResult.dart';
 import 'package:speech_to_text/speech_to_text.dart';
 import 'package:speech_to_text/speech_recognition_error.dart';
 import 'package:speech_to_text/speech_recognition_result.dart';
